Refactor test package structure and update Docker Compose paths
