+++ middleware/auth_test.go
@@ -2,7 +2,7 @@ package middleware
 
 import (
 	"testing"
-	test "the-wedding-game-api/_test"
+	test "the-wedding-game-api/_tests"
 	"the-wedding-game-api/db"
 	apperrors "the-wedding-game-api/errors"
 	"the-wedding-game-api/models"
